+++ packages/cross-chain/src/protocols/lifi-intents/provider.ts
@@ -24,8 +24,8 @@ import {
     LifiIntentsOrderStatusResponseSchema,
     LifiIntentsProviderConfigSchema,
     LifiIntentsQuoteResponseSchema,
-    LifiIntentsRoutesResponseSchema,
 } from "./schemas.js";
+import { parseRoutesIntoAssets } from "./services/parseRoutes.js";
 
 export class LifiIntentsProvider extends CrossChainProvider {
     static readonly PROTOCOL_NAME = "lifi-intents" as const;
@@ -222,40 +222,6 @@ export class LifiIntentsProvider extends CrossChainProvider {
     }
 
     private static parseRoutesResponse(data: unknown): NetworkAssets[] {
-        const { routes } = LifiIntentsRoutesResponseSchema.parse(data);
-        const chainAssetMap = new Map<
-            number,
-            Map<string, { address: string; symbol: string; decimals: number }>
-        >();
-
-        for (const route of routes) {
-            const fromChainId = Number(route.fromChain.chainId);
-            const toChainId = Number(route.toChain.chainId);
-            addToChainAssetMap(chainAssetMap, fromChainId, route.fromToken);
-            addToChainAssetMap(chainAssetMap, toChainId, route.toToken);
-        }
-
-        return Array.from(chainAssetMap.entries()).map(([chainId, assets]) => ({
-            chainId,
-            assets: Array.from(assets.values()),
-        }));
-    }
-}
-
-function addToChainAssetMap(
-    map: Map<number, Map<string, { address: string; symbol: string; decimals: number }>>,
-    chainId: number,
-    token: { address: string; symbol: string | null; name: string | null; decimals: number },
-): void {
-    if (!map.has(chainId)) {
-        map.set(chainId, new Map());
-    }
-    const key = token.address.toLowerCase();
-    if (!map.get(chainId)!.has(key)) {
-        map.get(chainId)!.set(key, {
-            address: token.address,
-            symbol: token.symbol ?? "",
-            decimals: token.decimals,
-        });
+        return parseRoutesIntoAssets(data);
     }
 }
