+++ packages/cross-chain/src/protocols/lifi-intents/services/LifiIntentsAssetDiscoveryService.ts
@@ -6,7 +6,7 @@ import {
     BaseAssetDiscoveryService,
     BaseAssetDiscoveryServiceConfig,
 } from "../../../core/services/BaseAssetDiscoveryService.js";
-import { LifiIntentsRoutesResponseSchema } from "../schemas.js";
+import { parseRoutesIntoAssets } from "./parseRoutes.js";
 
 export interface LifiIntentsAssetDiscoveryServiceConfig extends BaseAssetDiscoveryServiceConfig {
     orderServerUrl: string;
@@ -39,40 +39,6 @@ export class LifiIntentsAssetDiscoveryService extends BaseAssetDiscoveryService 
             );
         }
 
-        const { routes } = LifiIntentsRoutesResponseSchema.parse(response.data);
-        const chainAssetMap = new Map<
-            number,
-            Map<string, { address: string; symbol: string; decimals: number }>
-        >();
-
-        for (const route of routes) {
-            const fromChainId = Number(route.fromChain.chainId);
-            const toChainId = Number(route.toChain.chainId);
-            addToMap(chainAssetMap, fromChainId, route.fromToken);
-            addToMap(chainAssetMap, toChainId, route.toToken);
-        }
-
-        return Array.from(chainAssetMap.entries()).map(([chainId, assets]) => ({
-            chainId,
-            assets: Array.from(assets.values()),
-        }));
-    }
-}
-
-function addToMap(
-    map: Map<number, Map<string, { address: string; symbol: string; decimals: number }>>,
-    chainId: number,
-    token: { address: string; symbol: string | null; decimals: number },
-): void {
-    if (!map.has(chainId)) {
-        map.set(chainId, new Map());
-    }
-    const key = token.address.toLowerCase();
-    if (!map.get(chainId)!.has(key)) {
-        map.get(chainId)!.set(key, {
-            address: token.address,
-            symbol: token.symbol ?? "",
-            decimals: token.decimals,
-        });
+        return parseRoutesIntoAssets(response.data);
     }
 }
