+++ packages/cross-chain/src/protocols/lifi-intents/services/index.ts
@@ -1,2 +1,3 @@
 export { LifiIntentsAssetDiscoveryService } from "./LifiIntentsAssetDiscoveryService.js";
 export type { LifiIntentsAssetDiscoveryServiceConfig } from "./LifiIntentsAssetDiscoveryService.js";
+export { parseRoutesIntoAssets } from "./parseRoutes.js";
